+++ cdl/tests/features/images/contour_app.py
@@ -12,6 +12,7 @@ Contour finding application test
 
 import cdl.obj
 import cdl.param
+from cdl.env import execenv
 from cdl.tests import take_plotwidget_screenshot, test_cdl_app_context
 from cdl.tests.data import create_multigauss_image
 from cdl.tests.features.common.roi_app import create_test_image_with_roi
@@ -20,17 +21,23 @@ from cdl.tests.features.common.roi_app import create_test_image_with_roi
 def test():
     """Run contour finding application test scenario"""
     newparam = cdl.obj.new_image_param(height=200, width=200)
+    shapes = ("polygon", "circle", "ellipse")
     with test_cdl_app_context() as win:
         panel = win.imagepanel
-        ima1 = create_multigauss_image(newparam)
-        panel.add_object(ima1)
-        param = cdl.param.ContourShapeParam()
-        panel.processor.compute_contour_shape(param)
-        take_plotwidget_screenshot(panel, "contour_test")
-        ima2 = create_test_image_with_roi(newparam)
-        panel.add_object(ima2)
-        panel.processor.edit_regions_of_interest()
-        panel.processor.compute_contour_shape(param)
+        for shape in shapes:
+            ima1 = create_multigauss_image(newparam)
+            ima1.metadata["colormap"] = "gray"
+            panel.add_object(ima1)
+            param = cdl.param.ContourShapeParam.create(shape=shape)
+            panel.processor.compute_contour_shape(param)
+            take_plotwidget_screenshot(panel, "contour_test")
+            ima2 = create_test_image_with_roi(newparam)
+            ima2.metadata["colormap"] = "gray"
+            panel.add_object(ima2)
+            panel.processor.edit_regions_of_interest()
+            panel.processor.compute_contour_shape(param)
+            if not execenv.unattended:
+                break
 
 
 if __name__ == "__main__":